Stop leaving temp files after unit test runs
test_shell.CLOUD_2 is using an absolute path for a temp
file, so leaves /tmp/test_log_file around after the unit
tests are run.  Use a fixture instead so it's cleaned
automatically, which also removes the possibility of two
tests using the same file and interfering with each other.

Change-Id: If722b860be4010b91635c6d46f634da980e17152