Add server event command documentation for compute API 2.21
The 2.21 compute API microversion allows listing instance
action events and getting action event details for a deleted
server (which can be useful for auditing until the deleted
server is purged). As far as OSC is concerned it's just a
matter of specifying --os-compute-api-version 2.21 or higher
when listing events or showing event details, so this change
mentions 2.21 in the description of those commands.

Related to nova blueprint os-instance-actions-read-deleted-instances

Change-Id: If276c794f448b6fa5b0845499f3507a159acab85