Produce complete content for plugin docs
Plugin documentation previously used ``.. list-plugins::`` for each
plugin command, which just produced the summary line. This might be
useful if there were also a link to the complete (per-project) docs for
the commands; but since we have the content available, we might as well
produce the complete plugin docs inline. That's going to be most useful
to a reader anyway.

So this commit switches from ``list-plugins`` to ``autoprogram-cliff``
-- except for the following, whose docs break the build when this is
done:

- octavia
- rsd
- trove
- zun

These fixups are tracked under task #37241 under this same story.

Change-Id: I2f17e203fe3da92a709884c9052c8e39ff87f4c8
Story: #1735016
Task: #37240