Add plugin doc page for watcher
This was being omitted because whenever the plugin page was produced,
python-watcherclient wasn't in global-requirements. It is now, so
include a page for it in the plugin docs.

NOTE: We would like to use autoprogram-cliff to make the documentation
complete, but that breaks the build. For now, this is better than
nothing.

Change-Id: I49822242b9a0c031a053d6c2fd9f644a585f4ba5