Link to (some) plugin doc pages
We would like to use autoprogram-cliff to generate full docs inline for
each plugin. But for the following projects, that breaks the build:

- octavia
- rsd
- trove
- watcher
- zun

For those projects, we're using list-plugins instead, because that
builds, and it's better than nothing; but it only provides summaries of
the commands.

So with this commit, we add a link to the plugin documentation from the
actual plugin project where such documentation exists, which currently
is just:

- octavia
- watcher
- zun

(For rsd, I couldn't find openstack-published docs at all; for trove,
published docs exist, but the osc plugin isn't documented.)

Change-Id: I7c826ecef4319bead239e11b5f975302b2f24d1b
Story: #1735016
Task: #37244