Update a stale doc reference to use :neutron-doc:
The help page for network auto allocated topology had a link to a
newton-era networking guide document that has been superseded and is
now maintained in the neutron repository.

This commit adds 'neutron' to the openstackdocstheme configuration so
that the :neutron-doc: role works, and updates the link to point to the
modern version therein.

Change-Id: I5bcb40e265b22f15ff2f5ca4936160e231bb4075