Stop testing python 2 in tox and zuul.
Remove python 2 from envlist parameter. Check the link:

https://etherpad.openstack.org/p/drop-python2-support

The plan is drop the python 2 support from OpenStack in Ussuri release.
Remove the zuul jobs.

Note that the (non-voting) openstackclient-check-plugins job is still
running under py2 at this time. That will need to be fixed in the
python/openstackclient repository where the job is defined.

Change-Id: I3148db053b9ef0fcf7dc88e5cc075d974c93d819